+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [3.4.1] - 2026-04-14
+
+### Fixed
+- config.txt location was set to root folder instead of SK folder
+- download button worked, but save to SD was only saving to project folder not to the SD card
+- changes applie dto all appropriate files where the function is called
+
 ## [3.4.0] - 2026-03-27
 
 ### Added